refactor: split CLIcore_script_expand.c into focused sub-modules
CLIcore_script_expand.c was a 2,321-line file mixing four
unrelated concerns: arithmetic expression parsing, @-property
expansion, test condition evaluation, and environment variable
expansion.

Split into four files + one internal header:

  CLIcore_script_expand_internal.h
      ArithParser struct typedef shared by both
      expand_arith and expand_fps modules.

  CLIcore_script_expand_arith.c     (429 lines)
      Recursive-descent arithmetic parser:
        arith_skip_ws, arith_atom, arith_factor,
        arith_term, arith_shift, arith_compare,
        arith_bitwise_and, arith_bitwise_xor,
        arith_bitwise_or, arith_expr
      Shell expansion driver: cli_expand_arith()

  CLIcore_script_expand_fps.c      (1,189 lines)
      All @ property expanders:
        expand_fpsvar_write (write path handler)
        expand_fpsvar_seq   (@seq.NAME.prop)
        expand_fpsvar_procinfo/@proc namespace
        expand_fpsvar_stream/@s/@stream namespace
        expand_fpsvar_fps_strict/@fps namespace
        cli_expand_fpsvar() — main entry point

  CLIcore_script_expand_test.c      (578 lines)
      Bash-style test evaluator:
        test_unary_file (filesystem operators)
        test_unary_shm  (-S/-F/-P/-R/-n/-z/-v)
        test_binary_op  (-eq/-ne/-lt etc., ==/!=)
        cli_eval_test() — top-level evaluator

  CLIcore_script_expand_env.c       (620 lines)
      Variable expansion:
        emit_str_local helper
        cli_expand_env() — , ${VAR},
          array indexing, substring modifiers

CLIcore_script_expand.c retained as a stub comment
(compiles to zero symbols). All new files added to
MILKSCRIPT_SOURCES in CMakeLists.txt.

Verified: 276/276 CLI robustness tests pass.
